web/requirements.txt
django==4.0.4
djangorestframework==3.13.1
celery==5.2.6
redis==4.3.1
psycopg2-binary==2.9.3
sqlalchemy==1.4.36
pypandoc==1.8.1